add routerport instead of overloading device_id
fixes XXXX

Previously the l3_agent relied on overloading the device_id attribute to
link a port to a router. Overloading this attribute prevents using the
L3 extension in situations where the router might be a service VM (Nova
will not use a port that already has a device_id).

Change-Id: I23a3a45665db8f61ce94f624f52a019b89e51570