fix remove interface logic and update_ip logic
Previously, the code would only update remove an interface if the subnet
was the first one found by the db.  This fix searches all fixed_ips for
a matching subnet.

Additionally, this change also updates the update_port logic to enable a
user to update the IPs associated with router_port.

Change-Id: I4b86c39c574ac12bead918b1a288075bd58610a4